+++ src/main/java/com/yoo/lms/service/HomeworkBoardService.java
@@ -10,6 +10,8 @@ import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -22,32 +24,16 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 @Transactional(readOnly = true)
 @Slf4j
-@RequiredArgsConstructor
 public class HomeworkBoardService {
 
-    private HomeworkBoardRepository homeworkBoardRepository;
-    private CourseMaterialRepository courseMaterialRepository;
-    private CourseRepository courseRepository;
-    private StudentRepository studentRepository;
-    private EntityManager em;
-    private String baseDirectory;
-
-    @Autowired
-    public HomeworkBoardService(HomeworkBoardRepository homeworkBoardRepository,
-                                CourseMaterialRepository courseMaterialRepository,
-                                CourseRepository courseRepository,
-                                StudentRepository studentRepository,
-                                EntityManager em) {
-
-        this.homeworkBoardRepository = homeworkBoardRepository;
-        this.courseMaterialRepository = courseMaterialRepository;
-        this.courseRepository = courseRepository;
-        this.studentRepository = studentRepository;
-        this.em = em;
-        this.baseDirectory = "C:\\Users\\yoo-pc\\Desktop\\uploadedDirectory\\homeworkBoard\\";
-    }
+    private final HomeworkBoardRepository homeworkBoardRepository;
+    private final CourseMaterialRepository courseMaterialRepository;
+    private final CourseRepository courseRepository;
+    private final ViewRepository viewRepository;
+    private final String baseDirectory = "C:\\Users\\yoo-pc\\Desktop\\uploadedDirectory\\homeworkBoard\\";
 
 
     @Transactional
@@ -55,21 +41,13 @@ public class HomeworkBoardService {
                              Long courseId,
                              String title,
                              String content,
-                             Member member
-    ) throws IOException {
+                             Member member) throws IOException {
 
         Course course = courseRepository.findById(courseId).get();
-
-        // session에서 받아와야 함--
-
-//        Student student = studentRepository.findById("studentId1").get();
-
-        //---
 
         HomeworkBoard homeworkBoard = new HomeworkBoard(course, title, content, member);
 
         homeworkBoardRepository.save(homeworkBoard);
-        em.flush();
 
 
         // 파일이 없으면 게시물만 save하고 종료
@@ -95,11 +73,21 @@ public class HomeworkBoardService {
     }
 
     @Transactional
+    public void addViewCount(Long boardId){
+        HomeworkBoard homeworkBoard = homeworkBoardRepository.findById(boardId).get();
+
+        View view = new View(homeworkBoard);
+
+        viewRepository.save(view);
+        view.addViewCount(homeworkBoard);
+
+    }
+
+    @Transactional
     public void updateHomework(Long boardId,
                                String title,
                                String content,
-                               MultipartFile[] files
-    ) throws IOException {
+                               MultipartFile[] files) throws IOException {
 
 
         Optional<HomeworkBoard> homeworkBoardOptional = homeworkBoardRepository.findById(boardId);
@@ -121,10 +109,9 @@ public class HomeworkBoardService {
 
             File deletedFile = new File(material.getDirectory()+material.getFilename());
 
-            if(deletedFile.exists()) {
-                log.info("file exist!!!");
+            if(deletedFile.exists())
                 deletedFile.delete();
-            }
+
         }
 
         courseMaterialRepository.deleteAllByBoardId(boardId);
@@ -132,9 +119,9 @@ public class HomeworkBoardService {
         String savedDirectory = baseDirectory+homeworkBoard.getId() + "\\";
 
         File directory = new File(savedDirectory);
-        if(!directory.exists()) {
+        if(!directory.exists())
             directory.mkdirs();
-        }
+
 
         for (MultipartFile file : files) {
 
@@ -143,10 +130,6 @@ public class HomeworkBoardService {
 
             courseMaterialRepository.save(new CourseMaterial(homeworkBoard, fileName, savedDirectory));
         }
-
-
-
-
     }
 
     public HomeworkBoard findPostingById(Long boardId, boolean hasReplies) {
@@ -156,6 +139,7 @@ public class HomeworkBoardService {
         return homeworkBoard;
     }
 
+    @Transactional
     public void deleteByBoardId(Long boardId) {
 
         List<CourseMaterial> materials = courseMaterialRepository.findByBoardId(boardId);
@@ -164,40 +148,30 @@ public class HomeworkBoardService {
 
             File deletedFile = new File(material.getDirectory()+material.getFilename());
 
-            if(deletedFile.exists()) {
+            if(deletedFile.exists())
                 deletedFile.delete();
-            }
+
         }
 
-        File directory = new File(materials.get(0).getDirectory());
+        if(materials.size() > 0) {
 
-        if(directory.exists())
-            directory.delete();
+            File directory = new File(materials.get(0).getDirectory());
+
+            if (directory.exists())
+                directory.delete();
+        }
+
+        viewRepository.deleteAllByBoardId(boardId);
 
         courseMaterialRepository.deleteAllByBoardId(boardId);
-        homeworkBoardRepository.deletePostingById(boardId);
-
+        homeworkBoardRepository.deleteById(boardId);
     }
 
-    public List<BoardListDto> searchPosting(BoardSearchCondition condition, int page, int size) {
-        return homeworkBoardRepository.searchPosting(condition, page, size);
+    public String findTitleByBoardId(Long boardId) {
+        return homeworkBoardRepository.findTitleByBoardId(boardId);
     }
 
-    public long countTotalPosting(BoardSearchCondition condition, int page, int size, int numCurrentPageContent) {
-        return homeworkBoardRepository.countTotalPosting(condition, page, size, numCurrentPageContent);
-    }
-
-    public List<BoardListDto> searchPostingAllCriteria(BoardSearchCondition condition, int page, int size) {
-        return homeworkBoardRepository.searchPostingAllCriteria(condition, page, size);
-    }
-
-    public long countTotalAllCriteria(BoardSearchCondition condition, int page, int size, int numCurrentPageContent) {
-        return homeworkBoardRepository.countTotalAllCriteria(condition, page, size, numCurrentPageContent);
-    }
-
-    public PageMaker makePageMaker(int page, long totalCount, BoardSearchCriteria searchCriteria) {
-
-        return new PageMaker(page, totalCount, searchCriteria);
-
+    public Page<BoardListDto> searchPosting(BoardSearchCondition condition, boolean isMultipleCriteria, int page, int size) {
+        return homeworkBoardRepository.searchPosting(condition, isMultipleCriteria, PageRequest.of(page, size));
     }
 }
