+++ src/main/java/com/yoo/lms/service/QuestionBoardService.java
@@ -3,23 +3,16 @@ package com.yoo.lms.service;
 
 import com.yoo.lms.domain.*;
 import com.yoo.lms.dto.BoardListDto;
-import com.yoo.lms.repository.CourseMaterialRepository;
-import com.yoo.lms.repository.CourseRepository;
-import com.yoo.lms.repository.QuestionBoardRepository;
-import com.yoo.lms.repository.StudentRepository;
+import com.yoo.lms.repository.*;
 import com.yoo.lms.searchCondition.BoardSearchCondition;
-import com.yoo.lms.searchType.BoardSearchCriteria;
-import com.yoo.lms.tools.PageMaker;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.ui.Model;
 import org.springframework.web.multipart.MultipartFile;
-import org.springframework.web.multipart.MultipartHttpServletRequest;
 
-import javax.persistence.EntityManager;
-import javax.servlet.http.HttpServletRequest;
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
@@ -27,77 +20,33 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 @Transactional(readOnly = true)
 @Slf4j
 public class QuestionBoardService {
 
 
-    private QuestionBoardRepository questionBoardRepository;
-    private CourseMaterialRepository courseMaterialRepository;
-    private CourseRepository courseRepository;
-    private StudentRepository studentRepository;
-    private EntityManager em;
-    private String baseDirectory;
+    private final QuestionBoardRepository questionBoardRepository;
+    private final CourseMaterialRepository courseMaterialRepository;
+    private final CourseRepository courseRepository;
+    private final ViewRepository viewRepository;
+    private final String baseDirectory = "C:\\Users\\yoo-pc\\Desktop\\uploadedDirectory\\questionBoard\\";;
 
-
-    @Autowired
-    public QuestionBoardService(QuestionBoardRepository questionBoardRepository,
-                                CourseMaterialRepository courseMaterialRepository,
-                                CourseRepository courseRepository,
-                                StudentRepository studentRepository,
-                                EntityManager em) {
-
-        this.questionBoardRepository = questionBoardRepository;
-        this.courseMaterialRepository = courseMaterialRepository;
-        this.courseRepository = courseRepository;
-        this.studentRepository = studentRepository;
-        this.em = em;
-        this.baseDirectory = "C:\\Users\\yoo-pc\\Desktop\\uploadedDirectory\\questionBoard\\";
-    }
-
-    @Transactional
-    public void save(Long courseId,
-                     String title,
-                     String content
-                     ) {
-
-        Course course = courseRepository.findById(courseId).get();
-
-        //--- session에서 받아와야 함
-
-        Student student = studentRepository.findById("studentId1").get();
-
-        //---
-
-        QuestionBoard questionBoard = new QuestionBoard(course, title, content, student);
-
-        questionBoardRepository.save(questionBoard);
-    }
 
     @Transactional
     public void saveQuestion(MultipartFile[] files,
                      Long courseId,
                      String title,
                      String content,
-                     Member member
-    ) throws IOException {
+                     Member member) throws IOException {
 
         Course course = courseRepository.findById(courseId).get();
-
-        // session에서 받아와야 함--
-
-//        Student student = studentRepository.findById("studentId1").get();
-
-        //---
 
         QuestionBoard questionBoard = new QuestionBoard(course, title, content, member);
 
         questionBoardRepository.save(questionBoard);
-        em.flush();
-
 
         // 파일이 없으면 게시물만 save하고 종료
-        
         if(files[0].getOriginalFilename().equals(""))
             return;
 
@@ -105,9 +54,9 @@ public class QuestionBoardService {
 
         File directory = new File(savedDirectory);
 
-        if(!directory.exists()) {
+        if(!directory.exists())
             directory.mkdirs();
-        }
+
 
         for (MultipartFile file : files) {
 
@@ -116,7 +65,15 @@ public class QuestionBoardService {
 
             courseMaterialRepository.save(new CourseMaterial(questionBoard, fileName, savedDirectory));
         }
+    }
 
+    @Transactional
+    public void addViewCount(Long boardId){
+        QuestionBoard questionBoard = questionBoardRepository.findById(boardId).get();
+
+        View view = new View(questionBoard);
+        viewRepository.save(view);
+        view.addViewCount(questionBoard);
     }
 
     @Transactional
@@ -136,7 +93,6 @@ public class QuestionBoardService {
         findQuestion.updateInfo(title, content);
 
         // 파일이 없으면 게시물만 update하고 종료
-
         if(files[0].getOriginalFilename().equals(""))
             return;
 
@@ -146,10 +102,9 @@ public class QuestionBoardService {
 
             File deletedFile = new File(material.getDirectory()+material.getFilename());
 
-            if(deletedFile.exists()) {
-                log.info("file exist!!!");
+            if(deletedFile.exists())
                 deletedFile.delete();
-            }
+
         }
 
         courseMaterialRepository.deleteAllByBoardId(boardId);
@@ -169,18 +124,14 @@ public class QuestionBoardService {
             courseMaterialRepository.save(new CourseMaterial(findQuestion, fileName, savedDirectory));
         }
 
-
-
-
     }
 
     public QuestionBoard findPostingById(Long boardId) {
 
-        QuestionBoard questionBoard = questionBoardRepository.findPostingById(boardId);
-
-        return questionBoard;
+        return questionBoardRepository.findByIdFetchMember(boardId);
     }
 
+    @Transactional
     public void deleteByBoardId(Long boardId) {
 
         List<CourseMaterial> materials = courseMaterialRepository.findByBoardId(boardId);
@@ -194,38 +145,20 @@ public class QuestionBoardService {
             }
         }
 
-        File directory = new File(materials.get(0).getDirectory());
+        if(materials.size() > 0) {
+           File directory = new File(materials.get(0).getDirectory());
 
-        if(directory.exists())
-            directory.delete();
+            if (directory != null && directory.exists())
+                directory.delete();
+        }
+
+        viewRepository.deleteAllByBoardId(boardId);
 
         courseMaterialRepository.deleteAllByBoardId(boardId);
-        questionBoardRepository.deletePostingById(boardId);
-
+        questionBoardRepository.deleteById(boardId);
     }
 
-    public List<BoardListDto> searchPosting(BoardSearchCondition condition, int page, int size) {
-        return questionBoardRepository.searchPosting(condition, page, size);
+    public Page<BoardListDto> searchPosting(BoardSearchCondition condition, boolean isMultipleCriteria, int page, int size) {
+        return questionBoardRepository.searchPosting(condition, isMultipleCriteria, PageRequest.of(page, size));
     }
-
-    public long countTotalPosting(BoardSearchCondition condition, int page, int size, int numCurrentPageContent) {
-        return questionBoardRepository.countTotalPosting(condition, page, size, numCurrentPageContent);
-    }
-
-    public List<BoardListDto> searchPostingAllCriteria(BoardSearchCondition condition, int page, int size) {
-        return questionBoardRepository.searchPostingAllCriteria(condition, page, size);
-    }
-
-    public long countTotalAllCriteria(BoardSearchCondition condition, int page, int size, int numCurrentPageContent) {
-        return questionBoardRepository.countTotalAllCriteria(condition, page, size, numCurrentPageContent);
-    }
-
-    public PageMaker makePageMaker(int page, long totalCount, BoardSearchCriteria searchCriteria) {
-
-        return new PageMaker(page, totalCount, searchCriteria);
-
-    }
-
-
-
 }
